refactor dpm wave preview router
